+++ zaifapi/api_error.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 class ZaifApiError(Exception):
     def __init__(self, message):
         self._message = message
